CLMA_0003-iv

+++ UoR_MTS-Backend/mail-tracking-system/src/main/java/com/UoR_MTS_Backend/mail_tracking_system/model/MailRecord.java
@@ -18,7 +18,7 @@ import java.time.LocalDateTime;
         @GeneratedValue(strategy = GenerationType.IDENTITY)
         private int mailRecordId;
 
-        private long branchCode;
+        private int branchCode;
         private String branchName;
         private String mailType;
         private String trackingNumber;
